Data.Neither

Description

In this module, instances have been annotated with the interpretation of the Neither type that they use.

Documentation

data Neither a b

Constructors

Neither

neither :: a -> b -> Neither a b

Produces a Neither value whose types match the arguments, discarding the arguments